Add 'default' to the profile choices for a given network.
Add a new enum NetworkProfileChoice, listing the network profile choices
for a network. This is the set of network profiles plus 'default'.

When choosing the profile to activate, get the actual profile for each
network (so if it's default, use the *current* default profile) and, as
at present, select the least permissive.