When comparing network profile maps, ignore default entries.
Networks come and go. When they first appear, they get a profile type of
'default'. So if a config has a network with profile type 'default' and
another config does not have a record for that network, the configs are
the same. So when comparing network profile types, ignore all default
entries.

If we don't do that, we'll be invited to revert data when it's not
necessary.